LayoutTests/fast/css/getComputedStyle/getComputedStyle-outline-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("Test to make sure outline shorthand property returns CSSValueList properly.")

var testContainer = document.createElement("div");
testContainer.contentEditable = true;
document.body.appendChild(testContainer);

testContainer.innerHTML = '<div style="width:100px;height:100px"><div id="test">hello</div></div>';

e = document.getElementById('test');
computedStyle = window.getComputedStyle(e, null);

e.style.outline = "thick solid red";
shouldBe("computedStyle.getPropertyValue('outline')", "'5px solid rgb(255, 0, 0)'");
shouldBe("computedStyle.getPropertyCSSValue('outline').toString()", "'[object CSSValueList]'");
shouldBe("computedStyle.getPropertyCSSValue('outline').length", "3");
shouldBe("computedStyle.getPropertyCSSValue('outline').cssText", "'5px solid rgb(255, 0, 0)'");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(0).getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "5");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(1).getStringValue()", "'solid'");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "255");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "0");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "0");

e.style.outline = "2em solid";
shouldBe("computedStyle.getPropertyValue('outline')", "'32px solid rgb(0, 0, 0)'");
shouldBe("computedStyle.getPropertyCSSValue('outline').toString()", "'[object CSSValueList]'");
shouldBe("computedStyle.getPropertyCSSValue('outline').length", "3");
shouldBe("computedStyle.getPropertyCSSValue('outline').cssText", "'32px solid rgb(0, 0, 0)'");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(0).getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "32");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(1).getStringValue()", "'solid'");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "0");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "0");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "0");

e.style.outline = "4px none red";
shouldBe("computedStyle.getPropertyValue('outline')", "'4px none rgb(255, 0, 0)'");
shouldBe("computedStyle.getPropertyCSSValue('outline').toString()", "'[object CSSValueList]'");
shouldBe("computedStyle.getPropertyCSSValue('outline').length", "3");
shouldBe("computedStyle.getPropertyCSSValue('outline').cssText", "'4px none rgb(255, 0, 0)'");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(0).getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "4");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(1).getStringValue()", "'none'");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "255");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "0");
shouldBe("computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER)", "0");

document.body.removeChild(testContainer);

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
